1.6.0 Release. Go back to main stable Android Gradle plugin and fix major performance regression

+++ src/main/jni/interface/calc.h
@@ -62,7 +62,7 @@ typedef struct tagCALC {
 	char rom_version[32];
 	CalcModel model;
 
-	unsigned long long time_error;
+	uint64_t time_error;
 
 	BOOL active;
 	BOOL running;
